refactor(module): api -> fcfs-coupon-api rename
