+++ fcfs-coupon-api/src/main/java/com/coupop/fcfscoupon/FcfsCouponApplication.java
@@ -1,4 +1,4 @@
-package com.coupop;
+package com.coupop.fcfscoupon;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
