+++ fcfs-coupon-api/src/main/java/com/coupop/fcfscoupon/api/CouponController.java
@@ -1,10 +1,10 @@
-package com.coupop.api;
+package com.coupop.fcfscoupon.api;
 
-import com.coupop.api.dto.HistoryRequest;
-import com.coupop.api.dto.HistoryResponse;
-import com.coupop.api.dto.IssuanceRequest;
-import com.coupop.api.dto.ResendRequest;
-import com.coupop.api.support.RequestTime;
+import com.coupop.fcfscoupon.api.dto.HistoryRequest;
+import com.coupop.fcfscoupon.api.dto.HistoryResponse;
+import com.coupop.fcfscoupon.api.dto.IssuanceRequest;
+import com.coupop.fcfscoupon.api.dto.ResendRequest;
+import com.coupop.fcfscoupon.api.support.RequestTime;
 import com.coupop.fcfscoupon.common.exception.ApiException;
 import com.coupop.fcfscoupon.domain.coupon.CouponService;
 import com.coupop.fcfscoupon.domain.fcfs.FcfsIssueService;
