+++ fcfs-coupon-api/src/main/java/com/coupop/fcfscoupon/api/dto/ResendRequest.java
@@ -1,4 +1,4 @@
-package com.coupop.api.dto;
+package com.coupop.fcfscoupon.api.dto;
 
 public record ResendRequest(String historyId) {
 }
