+++ fcfs-coupon-api/src/main/java/com/coupop/fcfscoupon/api/support/RequestTime.java
@@ -1,4 +1,4 @@
-package com.coupop.api.support;
+package com.coupop.fcfscoupon.api.support;
 
 import java.time.LocalTime;
 import org.springframework.context.annotation.Scope;
