+++ fcfs-coupon-api/src/main/resources/application.properties
