+++ fcfs-coupon-api/src/test/java/com/coupop/fcfscoupon/FcfsCouponApplicationTests.java
@@ -1,4 +1,4 @@
-package com.coupop;
+package com.coupop.fcfscoupon;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
