+++ fcfs-coupon-api/src/test/java/com/coupop/fcfscoupon/api/testconfig/DatabaseSetUp.java
@@ -1,4 +1,4 @@
-package com.coupop.testconfig;
+package com.coupop.fcfscoupon.api.testconfig;
 
 import com.coupop.fcfscoupon.domain.fcfs.model.FcfsIssueRepository;
 import org.springframework.data.redis.core.StringRedisTemplate;
