+++ fcfs-coupon-api/src/test/java/com/coupop/fcfscoupon/api/testconfig/IntegrationTestConfig.java
@@ -1,10 +1,10 @@
-package com.coupop.testconfig;
+package com.coupop.fcfscoupon.api.testconfig;
 
 
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
 
-import com.coupop.api.support.RequestTime;
+import com.coupop.fcfscoupon.api.support.RequestTime;
 import com.coupop.fcfscoupon.domain.coupon.model.CouponEmailSender;
 import com.coupop.fcfscoupon.domain.coupon.model.RandomCodeGenerator;
 import com.coupop.fcfscoupon.domain.fcfs.model.FcfsIssuePolicy;
