+++ settings.gradle.kts
@@ -1,4 +1,5 @@
 rootProject.name = "fcfs-coupon"
-include("fcfs-coupon-domain-fcfs", "api")
+include("fcfs-coupon-api")
+include("fcfs-coupon-domain-fcfs")
 include("fcfs-coupon-domain-coupon")
 include("fcfs-coupon-common")
